refactor(typescript): client & hook to ts pt1 + imports
