+++ src/container/Hobbies/Hobbies.jsx
@@ -2,9 +2,9 @@ import React, { useState } from "react";
 import { HiChevronLeft, HiChevronRight } from "react-icons/hi";
 
 import { MotionWrap } from "../../wrapper";
-import { urlFor } from "../../client";
+import { urlFor } from "../../client.ts";
 import "./Hobbies.scss";
-import useFetch from "../../hooks/useFetch";
+import useFetch from "../../hooks/useFetch.ts";
 import Loader from "../../components/Loader/Loader";
 
 const Hobbies = () => {
